+++ families/creditcoin/ccprocessor-rust/.gitignore
@@ -1,3 +1,4 @@
 /target
 **/*.txt
 **/*.prof*
+**/*.html
