+++ families/creditcoin/ccprocessor-rust/Cargo.toml
@@ -25,7 +25,6 @@ sawtooth-sdk = { version = "0.5", path = "../sawtooth-sdk-rust" }
 serde_cbor = "0.11.1"
 serde_json = "1.0.64"
 sha2 = "0.9.5"
-shrinkwraprs = "0.3.0"
 typed-builder = "0.9.0"
 zmq = "0.9.2"
